feat: Implement Kanban board functionality with ticket management
- Add KanbanWorkspaceView component for managing boards and tickets.
- Create TicketCard component for displaying individual tickets.
- Develop Sidebar component for navigation and board management.
- Introduce SyncStatusBar for displaying synchronization status.
- Create TicketDetailPanel for viewing and editing ticket details.
- Implement routing for board pages with dynamic boardId.
- Add functionality for creating, deleting, and updating boards and tickets.
- Enhance user experience with context menus and command palette shortcuts.
