+++ src/lib/components/app/CommandPalette/CommandPalette.svelte
@@ -9,7 +9,7 @@
     } from "$lib/components/ui/dialog/index.js";
     import { Input } from "$lib/components/ui/input/index.js";
     import { Separator } from "$lib/components/ui/separator/index.js";
-    import type { CommandAction } from "./types.js";
+    import type { CommandAction } from "$lib/components/app/types.js";
 
     interface CommandPaletteProps {
         open: boolean;
