+++ src/lib/components/app/TicketDetail/TicketDetailPanel.svelte
@@ -2,48 +2,49 @@
     import { Button } from "$lib/components/ui/button/index.js";
     import { Separator } from "$lib/components/ui/separator/index.js";
     import { Sheet, SheetContent } from "$lib/components/ui/sheet/index.js";
-    import { useTicketDraft } from "$lib/hooks/ticket-draft.svelte";
     import TicketCommentsSection from "$lib/components/app/TicketDetail/TicketCommentsSection.svelte";
     import TicketDescriptionSection from "$lib/components/app/TicketDetail/TicketDescriptionSection.svelte";
     import TicketDetailHeader from "$lib/components/app/TicketDetail/TicketDetailHeader.svelte";
     import TicketEditorSection from "$lib/components/app/TicketDetail/TicketEditorSection.svelte";
-    import type { Ticket } from "./types.js";
+    import type { Ticket, TicketStatus } from "$lib/components/app/types.js";
 
     interface TicketDetailPanelProps {
         open: boolean;
         ticket: Ticket | null;
-        onUpdateTicket: (
-            ticketId: string,
-            updates: Partial<
-                Pick<Ticket, "title" | "description" | "status" | "labels">
-            >,
-        ) => void;
-        onAddComment: (ticketId: string, body: string) => void;
+        draftTitle: string;
+        draftDescription: string;
+        draftLabel: string;
+        draftStatus: TicketStatus;
+        statusOptions: Array<{ value: TicketStatus; label: string }>;
+        newCommentDraft: string;
+        onDraftTitleChange: (value: string) => void;
+        onDraftDescriptionChange: (value: string) => void;
+        onDraftLabelChange: (value: string) => void;
+        onDraftStatusChange: (value: TicketStatus) => void;
+        onNewCommentChange: (value: string) => void;
+        onSaveTicket: () => void;
+        onSubmitComment: () => void;
+        onEditorKeyDown: (event: KeyboardEvent) => void;
     }
 
     let {
         open = $bindable(false),
         ticket,
-        onUpdateTicket,
-        onAddComment,
+        draftTitle,
+        draftDescription,
+        draftLabel,
+        draftStatus,
+        statusOptions,
+        newCommentDraft,
+        onDraftTitleChange,
+        onDraftDescriptionChange,
+        onDraftLabelChange,
+        onDraftStatusChange,
+        onNewCommentChange,
+        onSaveTicket,
+        onSubmitComment,
+        onEditorKeyDown,
     }: TicketDetailPanelProps = $props();
-
-    const ticketDraft = useTicketDraft({
-        getTicket: () => ticket,
-        onUpdateTicket: (ticketId, updates) => {
-            onUpdateTicket(ticketId, updates);
-        },
-        onAddComment: (ticketId, body) => {
-            onAddComment(ticketId, body);
-        },
-    });
-
-    function onEditorKeyDown(event: KeyboardEvent) {
-        if ((event.metaKey || event.ctrlKey) && event.key === "Enter") {
-            event.preventDefault();
-            ticketDraft.saveTicket();
-        }
-    }
 </script>
 
 <Sheet bind:open>
@@ -58,25 +59,23 @@
 
                 <Separator />
 
-                <!-- svelte-ignore a11y_no_noninteractive_tabindex -->
-                <!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
                 <div
                     class="no-scrollbar min-h-0 flex-1 space-y-4 overflow-y-auto p-4"
                 >
                     <TicketEditorSection
-                        title={ticketDraft.draftTitle}
-                        label={ticketDraft.draftLabel}
-                        status={ticketDraft.draftStatus}
-                        statusOptions={ticketDraft.statusOptions}
-                        onTitleChange={ticketDraft.setDraftTitle}
-                        onLabelChange={ticketDraft.setDraftLabel}
-                        onStatusChange={ticketDraft.setDraftStatus}
+                        title={draftTitle}
+                        label={draftLabel}
+                        status={draftStatus}
+                        {statusOptions}
+                        onTitleChange={onDraftTitleChange}
+                        onLabelChange={onDraftLabelChange}
+                        onStatusChange={onDraftStatusChange}
                         {onEditorKeyDown}
                     />
 
                     <TicketDescriptionSection
-                        description={ticketDraft.draftDescription}
-                        onDescriptionChange={ticketDraft.setDraftDescription}
+                        description={draftDescription}
+                        onDescriptionChange={onDraftDescriptionChange}
                         {onEditorKeyDown}
                     />
 
@@ -84,9 +83,9 @@
 
                     <TicketCommentsSection
                         comments={ticket.comments}
-                        newCommentDraft={ticketDraft.newComment}
-                        onCommentChange={ticketDraft.setNewComment}
-                        onAddComment={ticketDraft.addComment}
+                        newCommentDraft={newCommentDraft}
+                        onCommentChange={onNewCommentChange}
+                        onAddComment={onSubmitComment}
                         {onEditorKeyDown}
                     />
                 </div>
@@ -97,9 +96,7 @@
                     <Button variant="outline" onclick={() => (open = false)}
                         >Close</Button
                     >
-                    <Button onclick={ticketDraft.saveTicket}
-                        >Save changes</Button
-                    >
+                    <Button onclick={onSaveTicket}>Save changes</Button>
                 </div>
             </div>
         {:else}
