+++ src/routes/(workspace)/board/+page.svelte
@@ -2,12 +2,9 @@
     <div
         class="w-full max-w-2xl rounded-xl border border-border/80 bg-surface-panel/80 p-6 text-center"
     >
-        <h1 class="font-display text-lg font-semibold">
-            Board View Coming Soon
-        </h1>
+        <h1 class="font-display text-lg font-semibold">Pick a Board</h1>
         <p class="mt-2 text-sm text-muted-foreground">
-            This route will host a focused per-board experience in a follow-up
-            step.
+            Use the sidebar to open a board-specific route.
         </p>
     </div>
 </div>
