+++ src/routes/(workspace)/board/[boardId]/+page.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+    import KanbanWorkspaceView from "$lib/components/app/Kanban/KanbanWorkspaceView.svelte";
+
+    interface BoardPageData {
+        boardId: string;
+    }
+
+    const { data }: { data: BoardPageData } = $props();
+</script>
+
+<KanbanWorkspaceView focusedBoardId={data.boardId} />
