+++ src/routes/(workspace)/board/[boardId]/+page.ts
@@ -1,0 +1,7 @@
+import type { PageLoad } from "./$types";
+
+export const load: PageLoad = async ({ params }) => {
+    return {
+        boardId: params.boardId,
+    };
+};
